pyproject.toml
[tool.poetry]
name = "strawberry-lambda-ws"
version = "0.1.0"
description = ""
authors = ["Patrick Arminio <patrick.arminio@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
strawberry-graphql = "^0.44.2"
mangum = "^0.10.0"
asgi-cors = "^0.3"

[tool.poetry.dev-dependencies]
black = {version = "^20.8b1", allow-prereleases = true}
flake8 = "^3.8.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
